[FEATURE / BUG FIX] Disable type feature on dates
This is to prevent improper data values inputted on the datepicker. Previous implementation has the typing feature enabled. Disabling the type feature will prevent the bug. This was fixed by adding a keyup function that would display the datepicker and letting the user to select any date he/she wanted.

Fixes: CRS reports bugs in Official Jan Ernest Test Cases Document