[ BUG FIX] Fix datepicker behaviour on changing the from date
Last Chosen Dates are Still Disabled even all the Date Inputs have been Cleared by the User

Fixed this by adding a blur function that would check for the value of the datepicker. If the value is empty or "", the maxDate(for datefrom) or the minDate(for dateto) will be reset to current date.

Fixes: #589