[ BUG FIX] Fix modal issue on clicking cancel or approve icon
Cancel Booking: A booking has to be cancelled twice before it changes its status. #613
Booking List: Approve, cancel and archive proceeds even if user is not yet confirming #600

To fix the previous bug for tooltip, icon was placed inside the <a>.

Fixed this by adding a condition which checks for the parent element's href. This will prevent the action click on the anchor tag. Action will proceed based on the answer that the user will click on the yes or no butotn.

Fixes: #600 #613