[ BUG FIX] Fix Navigation link on dispatch inland freight
Page Proceeds to Log In Module when the User Navigates to the Next Page of Dispatch Plan List #607

Previous implementation only lists item link as /viewFreightList.action.

Fixed this by adding a operations/viewInlandFreightList.action on the url item link. Fixed this for the whole inland freight dispatch plan. Limitation is that when ever you navigate away from a link, the selected link on the other tab's link is the selected tab you chose on the previous tab.

Fixes: #607